+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Arbitrary.hs
@@ -86,7 +86,6 @@ import Data.Text (pack)
 import Data.Word
 import Generic.Random (genericArbitraryU)
 import Numeric.Natural (Natural)
-import qualified PlutusLedgerApi.V1 as PV1
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Arbitrary (
   genValidAndUnknownCostModels,
@@ -95,22 +94,6 @@ import Test.Cardano.Ledger.Core.Arbitrary (
  )
 import Test.Cardano.Ledger.Mary.Arbitrary ()
 import Test.Cardano.Ledger.Plutus (alwaysFailsPlutus, alwaysSucceedsPlutus)
-
-instance Arbitrary PV1.Data where
-  arbitrary = resize 5 (sized gendata)
-    where
-      gendata n
-        | n > 0 =
-            oneof
-              [ PV1.I <$> arbitrary
-              , PV1.B <$> arbitrary
-              , PV1.Map <$> listOf ((,) <$> gendata (n `div` 2) <*> gendata (n `div` 2))
-              , PV1.Constr
-                  <$> fmap fromIntegral (arbitrary :: Gen Natural)
-                  <*> listOf (gendata (n `div` 2))
-              , PV1.List <$> listOf (gendata (n `div` 2))
-              ]
-      gendata _ = oneof [PV1.I <$> arbitrary, PV1.B <$> arbitrary]
 
 instance
   ( Arbitrary (AlonzoScript era)