+++ eras/alonzo/impl/CHANGELOG.md
@@ -31,6 +31,7 @@
 * Converted `CertState` to a type family
 * Expose `alonzoFixupFees`
 * Move `Arbitrary` instances for `Data`, `BinaryData` and `Datum` to `cardano-ledger-core`
+* Move `Arbitrary` instance for `PV1.Data` to `cardano-ledger-core`
 
 ## 1.12.0.0